requirements.txt
numpy<2.0.0
Pillow<10.3.0
requests<3.0.0
torch<=2.3.0
torchvision<=0.18.0
tqdm<5.0.0
matplotlib<4.0.0
faiss-cpu